requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
email-validator>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
aiofiles>=23.1.0
jinja2>=3.1.2
python-magic>=0.4.27 